hamlib: Use rigctld instead of cgo lib by default
Add support for the rigctld tcp daemon and default to using that instead of
linking against libhamlib (cgo).

This is more "go get" friendly as it does not require any external dependencies,
and it allows multiple applications to share the same rig.

Build with -tags libhamlib to link against the native C library.