wl2k: Add experimental gzip support
This commit adds experimental gzip support (in favor of lzhuf).

The feature is not enabled by default, but can be enabled by setting
GZIP_EXPERIMENT=1 at runtime.

Gzip compression is negotiated by adding character 'G' to SID.

If both parties supports gzip compression, then D-proposals is used
instead of the normal C-props (FC-lines).

D-proposals (FD-lines) have the same format as 'C's, but is used to
indicate that the data is gzip compressed.