Don't build with system installed static libax25
A new debian policy restricts the use of fPIC compiled static libraries,
which is (apparently) required by cgo. libax25 in Debian Stretch is
built without fPIC, so we can't use it anymore.

I've included a function in make.bash that downloads and compiles
libax25 to use when building the static Pat binary. Users can use the
env variable CGO_LDFLAGS to specify a custom libax25.a path.