Winlink CMS hostname changed to cms.winlink.org
This change was announced by the WDT after they migrated to the new
AWS 'cloud' CMS. The old hostname (server.winlink.org) is expected to
be operational for at least one year.

In addition to changing the config defaults, this commit includes a
(temporary) function that will rewrite user's config to replace the
old hostname.

Closes #104