feature request: send PTC commands on the command line
This is a feature request from https://github.com/la5nta/pat/issues/187
Idea is to send PTC commands on the command line. This is also useful
for bookmarks in the Web GUI.

Quite a number of changes were needed, as the commands need to be sent
to the PTC controller before going into WA8DED hostmode, so in the
init phase.  From the logic of Pat, the URL is passed to the
driver in the Dial phase. However, this is too late as the PTC is
already in host mode at that stage.

Changes to connect.go:

- in func Connect: check if an init parameter is in the URL and pass
them as type string to initPactorModem
- in func initPactorModem: add argument to take the init parameter and
pass it to pactor.OpenModem. Changes to the PACTOR modem are described
in the ptc-go repository https://github.com/harenber/ptc-go