Improved connect modal with rmslist
This also changes how the frequency input field is handled. Input
changes now triggers QSY commands right away, and the response is
interpreted to give the user a visual confirmation of the frequency
change. If the QSY fails, the field is ignored when building the final
connect URL so it won't cause an error when the user hits the connect
button.

Closes #186