Web GUI: Improved visibility of QSY errors
These changes should make it easier for users to notice if the QSY
command failed, before they hit the connect button. The connect button
will now be disabled as long as the frequency field is in focus,
preventing users from hitting the connect button before we're able to
perform the QSY and show an error to the user.

This is (hopefully) an improvement for the other QSY changes from
PR #210, following a discussion in #219.