Allow Unsorted Connect Aliases
Connect aliases are presented sorted into alphanumeric order, which is unnecessary and may be undesired. If the aliases are presented in the order they are entered in config.json then they can be in any order the user desires. This patch allows an unsorted ordering of connect aliases.

The re-ordering comes from storage in a Go map, which has no defined ordering. This patch uses  an array for the same purpose, which prevents re-ordering because arrays are inherently ordered.

In order to maintain backward compatibility in the format of config.json an optional new structure is allowed named unsorted_aliases of type [][]string. The first element of the subarray is the alias key and the second element is the URL. If entries are present in unsorted_aliases then they are presented in the aliases dropdown in unsorted order followed by any entries in connect_aliases sorted in alpha order.