Revert changes unrelated to Forms
Due to the Bootstrap and JQuery upgrade, the PR (la5nta/pat#234)
contained lots of unrelated changes that kept us from being able to
focus on the important bits needed to add support for Winlink Forms in
the web gui. This commit reverts all changes related to the upgrade of
those dependencies.

This breaks the nice accordion effect of the forms selection modal, but
the gist remains the same. Next up will be to rework the forms selection
without Bootstrap 4 features. Hopefully we'll be able to do so with a
good alternative UX design.