Configure a default forms path
This primes the user to run a forms update with zero configuration, but doesn't actually do it for them.  This could cause errors if they try to use forms before running the update. With some more work, we could make this more like the RMS list where the update happens proactively. For now, I think this is good enough.

Fixes #256