fix(manpage): update manpage to point users to proper paths for config, logs, and mailbox
e94f05f37cc5866f66e88d177dc39b2b5dc579a0 introduced a change, but the docs got missed. This follows up on that to fix the manpage documentation. As a chore at the same time, this fixes dev doc references to old travis-ci paths and adds the committer as a contributor.

Closes !326