forms: better parse support for to/cc/body
Add support for parsing To/Cc fields and passing an array
of addresses back to the compose window.

Improve handling of Msg. As per the templates [spec][]:

> Msg: -- Start of message body. All text following Msg: is placed in the body of the message.

and a few templates have text and variables on the same line as the "Msg:".
Therefore, simply strip out the Msg: and use that to mark the rest of
the template as being part of the body.

[spec]: http://www.philsherrod.com/Winlink/RMS_Express_Message_Templates.pdf

[Changed "var inBody = true" to "inBody := true" per Martin's
 suggestion]
Signed-off-by: Corey Minyard <cminyard@mvista.com>