Switch to DialURLContext
This adds generic support for aborting while dialing the connection. It
adds dial cancellation for ax25:// and telnet://.

The same mechanism is also used to cancel ARDOP dialing (implemented as
graceful disconnect), but keeps the old code to provide "dirty
disconnect" which is not supported through the context approach and/or
the generic conn.Close().