Switch to more idiomatic config fields for VARA
As VARA is hardcoded to use port n+1 for the data stream, it seems
unnecessary to support configuration of ctrl and data ports
individually. By removing this option, we can use the same address
config idioms as in the rest of the app. Also switched to from camelCase
to under_score as this is what we use for all other fields.

Implemented backwards compatible by implementing the json.Unmarshaller
interface.