Improve form variable prompting in CLI composer
Forms-enabled templates is not written with a CLI user in mind. By
echoing the template line before prompting for a <Var ...>, it will
hopefully be possible to understand the meaning of the variable (by
context).

Also apply a hack to refresh the regular expressions used by
varReplacer, to avoid prompting for the same variable multiple times.