Align auto-convert feature between web gui and cli
We have been auto-converting images for the web gui, while prompting the
user from the CLI composer. Given that the CLI prompt is causing issues
with the new "non-interactive" cli composer, I believe it's reasonable
to avoid this complexity by doing auto-convertion on the command line
aswell.

We could consider adding a config options to let the user control this
behavior in the future. But until then, let's keep it simple.

Resolves #431